refactor(www/audio): rename useMediaProgressQuery file
